+++ src/extension.ts
@@ -47,7 +47,8 @@ export async function activate(context: vscode.ExtensionContext) {
     })
   );
   // Linter is always activated but will only lint if compiler !== Disabled
-  const linter = new FortranLintingProvider(logger, context.storageUri.fsPath);
+  const linterCache = path.join(context.storageUri.fsPath);
+  const linter = new FortranLintingProvider(logger, linterCache);
   linter.activate(context);
   vscode.languages.registerCodeActionsProvider(FortranDocumentSelector(), linter);